+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to `laravel-livewire-wizard` will be documented in this file.
 
+## 1.1.2 - 2022-06-06
+
+### What's Changed
+
+- tests: fetch state for steps without state by @ju5t in https://github.com/spatie/laravel-livewire-wizard/pull/31
+
+**Full Changelog**: https://github.com/spatie/laravel-livewire-wizard/compare/1.1.1...1.1.2
+
+## 1.1.1 - 2022-06-03
+
+### What's Changed
+
+- fix: move EventEmitter to src by @ju5t in https://github.com/spatie/laravel-livewire-wizard/pull/30
+
+**Full Changelog**: https://github.com/spatie/laravel-livewire-wizard/compare/1.1.0...1.1.1
+
+## 1.1.0 - 2022-06-03
+
+### What's Changed
+
+- feat: make wizards testable by @ju5t in https://github.com/spatie/laravel-livewire-wizard/pull/29
+
+**Full Changelog**: https://github.com/spatie/laravel-livewire-wizard/compare/1.0.3...1.1.0
+
 ## 1.0.3 - 2022-05-30
 
 ### What's Changed